// frontend/src/enums/componentDerivedAttributes.ts
export enum ComponentDerivedAttributes {
  ComponentType = 'componentType',
  Width = 'componentWidth',
  Length = 'componentLength',
  Height = 'componentHeight',
  Volume = 'componentVolume',
  Weight = 'componentWeight',
  Count = 'componentCount',
  Rebar = 'componentRebar',
  MaterialPassport = 'materialPassport',
  LocationInRelationToBuilding = 'locationInRelationToBuilding'
}