// frontend/src/enums/componentKeyType.ts
//standardized attribute names for a component

export enum ComponentKeyType {
  Id = 'id',
  Img = 'img',
  Condition = 'condition',
  NoHarmfulSubstance = 'noHarmfulSubstance',
  AvailableFrom = 'availableFrom',
  Buyer = 'buyer',
  Price = 'price',
  PlanReference = 'planReference',
  LoadingCondition = 'loading_condition',
  Yaw = 'rotZAxis_yaw',
  Floor = 'floor',
  Location_x = 'location_x',
  Location_y = 'location_y',
}


//ReboundTestData = 'reboundTestData',
//VisualInspectionImages = 'visualInspectionImages',
//TypeOfElement = 'typeOfElement',
//RebarDiameterTop = 'rebarDiameterTop',
//RebarAmountTop = 'rebarAmountTop',
//RebarDiameterBottom = 'rebarDiameterBottom',
//RebarAmountBottom = 'rebarAmountBottom',
//Dimensions_l = 'dimensions_l',
 // Dimensions_w = 'dimensions_w',
 // Dimensions_h = 'dimensions_h',
 //Location_x = 'location_x',
 // Location_y = 'location_y',
 // Floor = 'floor',
 // Strength = 'strength',